[TASK] Move the uniqueness promise to `SlugGenerator`
This will allow us to reuse the slug generation in the `DataHandler`
if we need it as well.
