+++ Classes/Seo/SlugGenerator.php
@@ -6,7 +6,9 @@ namespace OliverKlee\Seminars\Seo;
 
 use Doctrine\DBAL\Driver\ResultStatement;
 use OliverKlee\Seminars\Domain\Model\Event\EventInterface;
+use TYPO3\CMS\Core\Database\Connection;
 use TYPO3\CMS\Core\Database\ConnectionPool;
+use TYPO3\CMS\Core\Database\Query\QueryBuilder;
 use TYPO3\CMS\Core\DataHandling\SlugHelper;
 use TYPO3\CMS\Core\Utility\GeneralUtility;
 
@@ -21,7 +23,7 @@ class SlugGenerator
     /**
      * @var non-empty-string
      */
-    private const TABLE_NAME = 'tx_seminars_seminars';
+    private const TABLE_NAME_EVENTS = 'tx_seminars_seminars';
 
     public function getPrefix(): string
     {
@@ -29,9 +31,7 @@ class SlugGenerator
     }
 
     /**
-     * Generates the slug for the given record.
-     *
-     * Note: This method does not check that the slug is unique.
+     * Generates a unique slug for the given record.
      *
      * @param array{record: array{title?: string, object_type?: string|int, topic?: string|int}} $parameters
      */
@@ -43,10 +43,7 @@ class SlugGenerator
 
         $title = '';
         if ($recordType === EventInterface::TYPE_EVENT_DATE) {
-            $queryBuilder = GeneralUtility::makeInstance(ConnectionPool::class)
-                ->getQueryBuilderForTable(self::TABLE_NAME);
-            $queryBuilder->getRestrictions()->removeAll();
-            $result = $queryBuilder->select('title')->from(self::TABLE_NAME)
+            $result = $this->getQueryBuilder()->select('title')->from(self::TABLE_NAME_EVENTS)
                 ->where('uid = :uid')->setParameter('uid', $topicUid)
                 ->execute();
             if ($result instanceof ResultStatement) {
@@ -65,6 +62,62 @@ class SlugGenerator
             $title = $record['title'] ?? '';
         }
 
-        return (new SlugHelper(self::TABLE_NAME, 'slug', []))->sanitize($title);
+        $slugCandidate = (new SlugHelper(self::TABLE_NAME_EVENTS, 'slug', []))->sanitize($title);
+
+        return $this->makeSlugUnique($slugCandidate);
+    }
+
+    private function getConnectionPool(): ConnectionPool
+    {
+        return GeneralUtility::makeInstance(ConnectionPool::class);
+    }
+
+    private function getQueryBuilder(): QueryBuilder
+    {
+        $queryBuilder = $this->getConnectionPool()->getQueryBuilderForTable(self::TABLE_NAME_EVENTS);
+        $queryBuilder->getRestrictions()->removeAll();
+
+        return $queryBuilder;
+    }
+
+    private function makeSlugUnique(string $slugCandidate): string
+    {
+        $slug = $slugCandidate;
+        $suffix = 0;
+
+        while ($this->countEventsWithSlug($slug) > 0) {
+            $suffix++;
+            $slug = $slugCandidate . '-' . $suffix;
+        }
+
+        return $slug;
+    }
+
+    private function countEventsWithSlug(string $slug): int
+    {
+        $queryBuilder = $this->getQueryBuilder();
+        $query = $queryBuilder
+            ->count('*')
+            ->from(self::TABLE_NAME_EVENTS)
+            ->where(
+                $queryBuilder->expr()->eq('slug', $queryBuilder->createNamedParameter($slug, Connection::PARAM_STR))
+            );
+
+        if (\method_exists($query, 'executeQuery')) {
+            $queryResult = $query->executeQuery();
+        } else {
+            $queryResult = $query->execute();
+        }
+        if ($queryResult instanceof ResultStatement) {
+            if (\method_exists($queryResult, 'fetchOne')) {
+                $count = (int)$queryResult->fetchOne();
+            } else {
+                $count = (int)$queryResult->fetchColumn(0);
+            }
+        } else {
+            $count = 0;
+        }
+
+        return $count;
     }
 }
