+++ Tests/Functional/Seo/Fixtures/SingleEventWithSlug.xml
@@ -1,0 +1,8 @@
+<?xml version="1.0" encoding="utf-8"?>
+<dataset>
+    <tx_seminars_seminars>
+        <uid>1</uid>
+        <title>Some event</title>
+        <slug>some-event</slug>
+    </tx_seminars_seminars>
+</dataset>
