+++ Tests/Functional/Seo/Fixtures/TwoSingleEventsWithSlug.xml
@@ -1,0 +1,13 @@
+<?xml version="1.0" encoding="utf-8"?>
+<dataset>
+    <tx_seminars_seminars>
+        <uid>1</uid>
+        <title>Some event</title>
+        <slug>some-event</slug>
+    </tx_seminars_seminars>
+    <tx_seminars_seminars>
+        <uid>2</uid>
+        <title>Some event</title>
+        <slug>some-event-1</slug>
+    </tx_seminars_seminars>
+</dataset>
